refactor: Change client and server
