+++ remote-wallet/client/wrc.go
@@ -1,4 +1,4 @@
-package client
+package main
 
 import (
 	"context"
